+++ projekatTSEO/src/main/java/projekatTSEO/repository/PohadjanjePredmetaRepository.java
@@ -9,7 +9,7 @@ import projekatTSEO.model.Predmet;
 import projekatTSEO.model.Profesor;
 import projekatTSEO.model.Student;
 
-public interface PohadjanjePredmetaRepository extends JpaRepository<PohadjanjePredmeta, Integer>{
+public interface PohadjanjePredmetaRepository extends JpaRepository<PohadjanjePredmeta, Long>{
 
 	List<PohadjanjePredmeta> findByStudent(Student s);
 	List<PohadjanjePredmeta> findByProfesor(Profesor s);
